requirements.txt
faiss-cpu==1.7.3
langchain==0.0.115
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.20.0
streamlit-chat==0.0.2.2